# Optimize file writing by using memory buffer.