--- a/docs/CONTRIBUTORS
+++ b/docs/CONTRIBUTORS
@@ -68,6 +68,7 @@ Lukasz Dalek
 Maarten De Braekeleer
 Maciej Sobkowski
 Marco Widmer
+Mariusz Glebocki
 Markus Krause
 Marlon James
 Marshal Qiao