@wsnyder 2022-06-04
Also note buffering will need to be done differently if using aio_write.