@mglb 2022-06-14
Total time measurements with a large project. Each average calculated from 8 runs.
<table>
  <thead>
  <tr>
    <td rowspan="2">Output storage type</td>
    <td rowspan="2">Allocator</td>
    <td colspan="2">master</td>
    <td colspan="2">PR</td>
    <td colspan="2">master - PR</td>
  </tr>
  <tr>
    <td>Average</td>
    <td>Median</td>
    <td>Average</td>
    <td>Median</td>
    <td>Average</td>
    <td>Median</td>
  </tr>
  </thead>
  <tbody>
    <tr> <td>HDD</td> <td>glibc</td>    <td>394.0</td> <td>393.9</td> <td>393.4</td> <td>393.1</td> <td>-0.5</td> <td>-0.7</td> </tr>
    <tr> <td>HDD</td> <td>tcmalloc</td> <td>249.7</td> <td>249.4</td> <td>246.0</td> <td>245.0</td> <td>-3.7</td> <td>-4.4</td> </tr>
    <tr> <td>SSD</td> <td>glibc</td>    <td>393.9</td> <td>294.4</td> <td>386.3</td> <td>385.5</td> <td>-7.6</td> <td>-8.9</td> </tr>
    <tr> <td>SSD</td> <td>tcmalloc</td> <td>247.6</td> <td>246.9</td> <td>243.1</td> <td>242.4</td> <td>-4.5</td> <td>-4.5</td> </tr>
  </tbody>
</table>