@gezalore 2022-06-14
How much of the speedup is due to aio and due to manual buffering?