@mglb 2022-06-15
> How much of the speedup is due to aio and due to manual buffering?

Most of the speedup is due to manual buffering. The PR initially used `fwrite()`, and the results were similar (see the description).
I think aio advantage is not visible in my tests because I'm testing it on a machine with 64GB of RAM with a project that has peak memory use around 22GB. Most of writes land in FS cache, and are synced after Verilator exits.
I'll try to run the test on a machine with 32GB of RAM.